+++ src/main/java/exception/CommandNotRecognisedException.java
@@ -1,4 +1,4 @@
-package duke.exception;
+package exception;
 
 public class CommandNotRecognisedException extends DukeException {
 }
