+++ src/main/java/exception/IllegalCharacterException.java
@@ -1,4 +1,4 @@
-package duke.exception;
+package exception;
 
 public class IllegalCharacterException extends DukeException {
 }
