+++ src/main/java/task/Task.java
@@ -1,4 +1,4 @@
-package duke.task;
+package task;
 
 public class Task {
     protected final String description;
